Fix warnings.

+++ indra/llcommon/llerror.cpp
@@ -569,7 +569,7 @@ namespace LLError
 		mTags(new const char*[tag_count]),
 		mTagCount(tag_count)
 	{
-		for (int i = 0; i < tag_count; i++)
+		for (size_t i = 0; i < tag_count; i++)
 		{
 			mTags[i] = tags[i];
 		}
